chore: move proto to separate top level package
