+++ .circleci/config.yml
@@ -120,6 +120,7 @@ commands:
       - restore_cache:
           keys:
             - det-go-deps-v1dev6-{{ checksum  "master/go.sum" }}-{{ checksum  "agent/go.sum" }}
+      - run: make -C proto get-deps
       - run: make -C master get-deps
       - run: make -C agent get-deps
       - save_cache:
@@ -694,6 +695,8 @@ jobs:
     steps:
       - checkout
       - go-get-deps
+      - run: make -C proto build
+      - run: make -C proto check
       - run: make -C master check
       - run: make -C agent check
 
@@ -705,6 +708,7 @@ jobs:
     steps:
       - checkout
       - go-get-deps
+      - run: make -C proto build
       - run: make -C master build
       - run: make -C agent build
       - persist_to_workspace:
@@ -721,6 +725,7 @@ jobs:
     steps:
       - checkout
       - go-get-deps
+      - run: make -C proto build
       - run: make -C master test
       - run: make -C agent test
