+++ Makefile
@@ -8,6 +8,7 @@ get-deps:
 	pip install -r requirements.txt
 	$(MAKE) -C master $@
 	$(MAKE) -C agent $@
+	$(MAKE) -C proto $@
 	$(MAKE) -C webui $@
 
 .PHONY: package
@@ -22,7 +23,7 @@ build-%:
 build-docs: build-common build-harness build-cli build-deploy build-examples
 	$(MAKE) -C docs build
 .PHONY: build-master
-build-master: build-docs build-webui-elm build-webui-react
+build-master: build-docs build-webui-elm build-webui-react build-proto
 	$(MAKE) -C master build
 .PHONY: build
 build: build-master build-agent
@@ -31,13 +32,13 @@ build: build-master build-agent
 clean-%:
 	$(MAKE) -C $(subst -,/,$*) clean
 .PHONY: clean
-clean: clean-tools clean-common clean-harness clean-cli clean-deploy clean-examples clean-docs clean-webui clean-master clean-agent
+clean: clean-tools clean-proto clean-common clean-harness clean-cli clean-deploy clean-examples clean-docs clean-webui clean-master clean-agent
 
 .PHONY: check-%
 check-%:
 	$(MAKE) -C $(subst -,/,$*) check
 .PHONY: check
-check: check-common check-harness check-cli check-deploy check-e2e_tests check-master check-agent check-webui check-examples
+check: check-common check-proto check-harness check-cli check-deploy check-e2e_tests check-master check-agent check-webui check-examples
 
 .PHONY: fmt-%
 fmt-%:
