+++ agent/go.mod
@@ -22,10 +22,12 @@ require (
 	github.com/spf13/cobra v1.0.0
 	github.com/spf13/pflag v1.0.5
 	golang.org/x/sys v0.0.0-20200509044756-6aff5f38e54f
-	golang.org/x/tools v0.0.0-20200502202811-ed308ab3e770
+	golang.org/x/tools v0.0.0-20200522201501-cb1345f3a375
 	gotest.tools v2.2.0+incompatible // indirect
 )
 
 replace github.com/determined-ai/determined/master => ../master
+
+replace github.com/determined-ai/determined/proto => ../proto
 
 replace github.com/docker/docker v1.13.1 => github.com/docker/engine v1.4.2-0.20191113042239-ea84732a7725
