+++ master/.gitignore
@@ -19,7 +19,3 @@ vendor/
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
-
-# Proto generated files
-pkg/proto
-static/swagger-spec/
