+++ master/.goreleaser.yml
@@ -3,7 +3,7 @@ project_name: determined-master
 before:
   hooks:
     - make pre-package
-    - make proto
+    - make -C ../proto build
 
 snapshot:
   name_template: "{{ .Tag }}"
