+++ proto/.gitignore
@@ -1,0 +1,3 @@
+# Proto generated files
+pkg/proto/
+build/
