+++ proto/go.mod
@@ -1,0 +1,10 @@
+module github.com/determined-ai/determined/proto
+
+go 1.13
+
+require (
+	github.com/bufbuild/buf v0.16.0
+	github.com/golang/protobuf v1.4.2
+	github.com/grpc-ecosystem/grpc-gateway v1.14.6
+	google.golang.org/grpc v1.29.1
+)
