+++ proto/src/agent/v1/agent.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 package agent.v1;
-option go_package = "github.com/determined-ai/determined/master/pkg/proto/agentv1";
+option go_package = "github.com/determined-ai/determined/proto/pkg/agentv1";
 
 import "google/protobuf/timestamp.proto";
