+++ proto/src/api/v1/agent.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 package api.v1;
-option go_package = "github.com/determined-ai/determined/master/pkg/proto/apiv1";
+option go_package = "github.com/determined-ai/determined/proto/pkg/apiv1";
 
 import "agent/v1/agent.proto";
