+++ proto/tools.go
@@ -1,0 +1,11 @@
+// +build tools
+
+package tools
+
+import (
+	_ "github.com/bufbuild/buf/cmd/buf"
+	_ "github.com/golang/protobuf/protoc-gen-go"
+	_ "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway"
+	_ "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger"
+	_ "google.golang.org/grpc"
+)
