+++ webui/tests/test-cluster/Makefile
@@ -24,6 +24,7 @@ prep-root:
 	@mkdir -p build/webui/react
 	@mkdir -p build/wheels/
 	@ln -s $(abspath ../../../master/static) build/
+	@ln -s $(abspath ../../../proto/build/swagger) build/swagger
 	@ln -s $(abspath ../../../docs/site/html/*) build/webui/docs
 	@ln -s $(abspath ../../../webui/react/build/*) build/webui/react
 	@ln -s $(abspath ../../../webui/elm/public) build/webui/elm
